remove piggy bank, organised code
