+++ src/main/java/net/fullstackjones/bigbraincurrency/BigBrainCurrency.java
@@ -1,12 +1,9 @@
 package net.fullstackjones.bigbraincurrency;
 
-import net.fullstackjones.bigbraincurrency.block.ModBlockEntities;
-import net.fullstackjones.bigbraincurrency.block.ModBlocks;
-import net.fullstackjones.bigbraincurrency.block.entities.ShopBlockEntityRenderer;
+import net.fullstackjones.bigbraincurrency.registration.*;
+import net.fullstackjones.bigbraincurrency.entities.ShopBlockEntityRenderer;
 import net.fullstackjones.bigbraincurrency.data.ModAttachmentTypes;
-import net.fullstackjones.bigbraincurrency.item.ModItems;
-import net.fullstackjones.bigbraincurrency.loottables.ModLootModifiers;
-import net.fullstackjones.bigbraincurrency.menu.ModContainers;
+import net.fullstackjones.bigbraincurrency.registration.ModMenus;
 import net.fullstackjones.bigbraincurrency.menu.MoneyPouchScreen;
 import net.fullstackjones.bigbraincurrency.menu.ShopScreen;
 import net.neoforged.neoforge.client.event.EntityRenderersEvent;
@@ -27,8 +24,8 @@ import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.neoforged.neoforge.common.NeoForge;
 import net.neoforged.neoforge.event.server.ServerStartingEvent;
 
-import static net.fullstackjones.bigbraincurrency.menu.ModContainers.MONEYPOUCHMENU;
-import static net.fullstackjones.bigbraincurrency.menu.ModContainers.SHOPMENU;
+import static net.fullstackjones.bigbraincurrency.registration.ModMenus.MONEYPOUCHMENU;
+import static net.fullstackjones.bigbraincurrency.registration.ModMenus.SHOPMENU;
 
 // The value here should match an entry in the META-INF/neoforge.mods.toml file
 @Mod(BigBrainCurrency.MODID)
@@ -53,7 +50,7 @@ public class BigBrainCurrency
         ModItems.register(modEventBus);
         ModBlockEntities.register(modEventBus);
         ModBlocks.register(modEventBus);
-        ModContainers.register(modEventBus);
+        ModMenus.register(modEventBus);
         ModLootModifiers.register(modEventBus);
 
     }
