+++ src/main/java/net/fullstackjones/bigbraincurrency/block/BrainBankBlock.java
@@ -1,0 +1,27 @@
+package net.fullstackjones.bigbraincurrency.block;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.EntityBlock;
+import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.entity.BlockEntityTicker;
+import net.minecraft.world.level.block.entity.BlockEntityType;
+import net.minecraft.world.level.block.state.BlockState;
+import org.jetbrains.annotations.Nullable;
+
+public class BrainBankBlock extends Block implements EntityBlock {
+    public BrainBankBlock(Properties props) {
+        super(props);
+    }
+
+    @Override
+    public @Nullable BlockEntity newBlockEntity(BlockPos pos, BlockState state) {
+        return null;
+    }
+
+    @Override
+    public @Nullable <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> blockEntityType) {
+        return EntityBlock.super.getTicker(level, state, blockEntityType);
+    }
+}
