+++ src/main/java/net/fullstackjones/bigbraincurrency/registration/ModBlockEntities.java
@@ -1,16 +1,10 @@
-package net.fullstackjones.bigbraincurrency.block;
+package net.fullstackjones.bigbraincurrency.registration;
 
 import net.fullstackjones.bigbraincurrency.BigBrainCurrency;
-import net.fullstackjones.bigbraincurrency.block.entities.ShopBlockEntity;
-import net.fullstackjones.bigbraincurrency.block.entities.ShopBlockEntityRenderer;
-import net.minecraft.client.renderer.blockentity.BlockEntityRenderers;
-import net.minecraft.core.Registry;
+import net.fullstackjones.bigbraincurrency.entities.ShopBlockEntity;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.world.level.block.entity.BlockEntityType;
-import net.neoforged.api.distmarker.Dist;
-import net.neoforged.api.distmarker.OnlyIn;
 import net.neoforged.bus.api.IEventBus;
-import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 import net.neoforged.neoforge.registries.DeferredRegister;
 
 import java.util.function.Supplier;
